Fix test with screen stack
